Revert "cambio dependencias"
This reverts commit 56f794220e8aabef6de029bdd2f1df4883bbddad.
