+++ src/main/java/database/DirectorDAO.java
@@ -1,5 +1,6 @@
-package model;
+package database;
 
+import model.Director;
 import java.util.List;
 
 public interface DirectorDAO {
