+++ src/main/java/database/EstudianteDAOJDBC.java
@@ -1,8 +1,6 @@
 package database;
 
 import model.Estudiante;
-import model.EstudianteDAO;
-
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
