+++ src/main/java/database/ProyectoDAO.java
@@ -1,0 +1,9 @@
+package database;
+
+import model.Proyecto;
+
+import java.sql.SQLException;
+
+public interface ProyectoDAO {
+    void guardar(Proyecto proyecto) throws SQLException;
+}
