+++ src/main/java/database/TutorDAO.java
@@ -1,5 +1,6 @@
-package model;
+package database;
 
+import model.Tutor;
 import java.util.List;
 
 public interface TutorDAO {
