+++ src/main/java/database/TutorDAOJDBC.java
@@ -1,8 +1,6 @@
 package database;
 
 import model.Tutor;
-import model.TutorDAO;
-
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
