--- src/main/java/model/AdministradorDAO.java
@@ -1,4 +1,0 @@
-package model;
-
-public interface AdministradorDAO {
-}
