--- src/main/java/model/ConvenioDAO.java
@@ -1,6 +1,0 @@
-package model;
-
-public interface ConvenioDAO {
-    void create(Convenio convenio);
-    void updateArchivo(Convenio convenio);
-}
