--- src/main/java/model/ProyectoDAO.java
@@ -1,8 +1,0 @@
-package model;
-
-import java.util.List;
-
-public interface ProyectoDAO {
-    void guardar(Proyecto proyecto);
-    public List<Proyecto> obtenerProyectos();
-}
