+++ src/main/java/org/example/gpps_unrn/ProyectoFormController.java
@@ -1,15 +1,12 @@
 package org.example.gpps_unrn;
 
-//import database.*;
-import database.Conn;
+import database.*;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.util.StringConverter;
 import java.sql.Connection;
-import java.sql.SQLException;
-
 import model.*;
 
 public class ProyectoFormController {
@@ -31,12 +28,7 @@ public class ProyectoFormController {
     public void initialize() {
         try {
             Connection conn = Conn.getConnection();
-            proyectoDAO = new ProyectoDAO() {
-                @Override
-                public void guardar(Proyecto proyecto) throws SQLException {
-
-                }
-            };
+            proyectoDAO = new ProyectoDAOJDBC(conn);
             estudianteDAO = new EstudianteDAOJDBC(conn);
             directorDAO = new DirectorDAOJDBC(conn);
             tutorDAO = new TutorDAOJDBC(conn);
