+++ src/test/java/ConvenioTest.java
@@ -1,5 +1,3 @@
-package model;
-
 import org.junit.jupiter.api.Test;
 
 public class ConvenioTest {
